Criação da tela de início + arquivo de inserções
