--- Projeto - Android Studio/app/src/main/java/com/example/harppia/HomeActivity.java
@@ -1,45 +1,0 @@
-package com.example.harppia;
-
-import android.os.Bundle;
-import android.view.View;
-import android.view.Window;
-import android.widget.Button;
-
-import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
-
-import com.google.android.material.transition.platform.MaterialSharedAxis;
-
-public class HomeActivity extends AppCompatActivity {
-
-    Button bt_vorta;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        getWindow().requestFeature(Window.FEATURE_ACTIVITY_TRANSITIONS);
-        getWindow().setEnterTransition(new MaterialSharedAxis(MaterialSharedAxis.X, true));
-        getWindow().setReturnTransition(new MaterialSharedAxis(MaterialSharedAxis.X, false));
-
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_home);
-
-        bt_vorta = findViewById(R.id.bt_vorta);
-
-        bt_vorta.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finishAfterTransition();
-            }
-        });
-
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
-            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
-            return insets;
-        });
-    }
-}
